# swagger/service/addons_v1.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "service/addons_v1.proto",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "StoreService"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/addons/section/{handle}": {
      "put": {
        "operationId": "StoreService_UpdateSection",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/store_serviceNoResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "handle",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/store_serviceUpdateSection"
            }
          }
        ],
        "tags": [
          "StoreService"
        ]
      }
    },
    "/v1/addons/sections": {
      "get": {
        "operationId": "StoreService_ListSections",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/store_serviceListSectionsResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "locale.lang",
            "in": "query",
            "required": false,
            "type": "string",
            "enum": [
              "UNSPECIFIED",
              "de",
              "en",
              "fr"
            ],
            "default": "UNSPECIFIED"
          },
          {
            "name": "locale.country",
            "in": "query",
            "required": false,
            "type": "string",
            "enum": [
              "UNDEFINED",
              "DE",
              "US"
            ],
            "default": "UNDEFINED"
          },
          {
            "name": "week.year",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "week.week",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int64"
          }
        ],
        "tags": [
          "StoreService"
        ]
      }
    }
  },
  "definitions": {
    "domainCountry": {
      "type": "string",
      "enum": [
        "UNDEFINED",
        "DE",
        "US"
      ],
      "default": "UNDEFINED"
    },
    "domainLanguage": {
      "type": "string",
      "enum": [
        "UNSPECIFIED",
        "de",
        "en",
        "fr"
      ],
      "default": "UNSPECIFIED"
    },
    "domainLocale": {
      "type": "object",
      "properties": {
        "lang": {
          "$ref": "#/definitions/domainLanguage"
        },
        "country": {
          "$ref": "#/definitions/domainCountry"
        }
      }
    },
    "domainSection": {
      "type": "object",
      "properties": {
        "handle": {
          "type": "string"
        },
        "index": {
          "type": "integer",
          "format": "int32"
        },
        "singular": {
          "type": "string"
        },
        "plural": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "imageUri": {
          "type": "string"
        }
      }
    },
    "domainYearWeek": {
      "type": "object",
      "properties": {
        "year": {
          "type": "integer",
          "format": "int64"
        },
        "week": {
          "type": "integer",
          "format": "int64"
        }
      }
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
    "store_serviceListSectionsResponse": {
      "type": "object",
      "properties": {
        "sections": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/domainSection"
          }
        }
      }
    },
    "store_serviceNoResponse": {
      "type": "object"
    },
    "store_serviceUpdateSection": {
      "type": "object",
      "properties": {
        "index": {
          "type": "integer",
          "format": "int64"
        },
        "description": {
          "type": "string"
        },
        "imageUri": {
          "type": "string"
        }
      }
    }
  }
}